+++ packages/bottender-facebook/src/FacebookClient.ts
@@ -1,7 +1,7 @@
 import AxiosError from 'axios-error';
 import get from 'lodash/get';
 import { MergeExclusive } from 'type-fest';
-import { MessengerClient } from 'bottender';
+import { MessengerClient } from '@vraksha/bottender';
 
 import * as Types from './FacebookTypes';