+++ packages/create-bottender-app/src/utils/generateAppEntry.ts
@@ -28,7 +28,7 @@ module.exports = async function App(context) {
 
   return prettier.format(
     `
-import { Action, ${contexts.join(', ')} } from 'bottender';
+import { Action, ${contexts.join(', ')} } from '@vraksha/bottender';
 
 export default async function App(context: ${contexts.join(
       ' | '